backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
pydantic>=2
python-dotenv
opencv-python-headless
numpy
python-multipart
requests
pandas
scikit-learn